% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmark.R
\name{Benchmark}
\alias{Benchmark}
\title{"Utility_Benchmark" :  For internal use only}
\usage{
Benchmark(Calc, Ref)
}
\arguments{
\item{Calc}{Computed matrix (0,1), to compare with reference}

\item{Ref}{Reference matrix (0,1)}
}
\value{
Vector		Returns a vector : TP, TN, FP, FN, Sensitivity, Specificity, Dist to the diagonal
}
\description{
This function is a utility intended to compute values in order to compare methods.
For internal use only.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{		Calc and Ref must have the same size and contain 0 or 1 only.
}\if{html}{\out{</div>}}
}
