% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRARegress.R
\name{CheckInputData}
\alias{CheckInputData}
\title{Check the input data for function MRARegress}
\usage{
CheckInputData(MatExp, Perturb, NodeName, KnlgMap, Method, MapExper)
}
\arguments{
\item{MatExp}{The "Expression Matrix".}

\item{Perturb}{Refers to the name of the perturbations and the nodes they act upon.}

\item{NodeName}{The name of the nodes.}

\item{KnlgMap}{The "Knowledge Map".}

\item{Method}{The method to use to compute the connectivity matrix.}

\item{MapExper}{The "Experience Map" matrix}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{		A message if an error is detected and returns NULL otherwise.
}\if{html}{\out{</div>}}
}
\description{
This function checks the input data for function MRARegress.
The parameters are same as those of MRARegress.
}
