% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCI.R
\name{CheckInputDataCI}
\alias{CheckInputDataCI}
\title{Checks the input data for function rCI}
\usage{
CheckInputDataCI(
  Ret,
  Func,
  St,
  P0,
  Perturbs,
  nbRep,
  nbD,
  K,
  Solution,
  Root,
  Verbose
)
}
\arguments{
\item{Ret}{list of informations delivered by "MRARegress".}

\item{Func}{Function of P (parameters) and X (expression values to compute).}

\item{St}{Multiroot starts from these values, to solve the system of ODE.}

\item{P0}{Initial value of the parameters to perturb.}

\item{Perturbs}{Set of perturbations.}

\item{nbRep}{Number of replicates to generate for every perturbation, including the initial value.}

\item{nbD}{Number of random draws to compute the "Confidence Intervals" ('CI').}

\item{K}{Used to compute the simulated noise.}

\item{Solution}{Name of a '.rda' file ('***/xxx.rda'), containing the exact values of the connectivity coefficients.}

\item{Root}{Name of a writable directory ('yyyy').}

\item{Verbose}{If TRUE, additional printings are made. These printings are for internal use only, so they are not documented.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{		A message if an error is detected and returns NULL otherwise.
}\if{html}{\out{</div>}}
}
\description{
This function checks the input data for function rCI.
The parameters are the same as those of rCI.
}
