% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classify.R
\name{CheckInputDataCY}
\alias{CheckInputDataCY}
\title{Checks the input data for function Classify}
\usage{
CheckInputDataCY(Ret, Classes, MethDiscr, Lbda, Verbose, NoPrint)
}
\arguments{
\item{Ret}{list of informations delivered by "MRARegress" or a matrix.}

\item{Classes}{A vector showing the classes to define.}

\item{MethDiscr}{The method to use to classify the matrix.}

\item{Lbda}{An hyper-parameter specific to the method.}

\item{Verbose}{If TRUE, additional printings are made. These printings are for internal use only, so they are not documented.}

\item{NoPrint}{If TRUE, no printings are made. Useful for tests including calls to 'MRARegress' in a loop.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{		A message if an error is detected and returns NULL otherwise.
}\if{html}{\out{</div>}}
}
\description{
This function checks the input data for function Classify.
The parameters are the same as those of Classify.
}
