% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawDiscr.R
\name{CheckInputDataDD}
\alias{CheckInputDataDD}
\title{Checks the input data for function DrawDiscr}
\usage{
CheckInputDataDD(rDiscr, Ref, Title)
}
\arguments{
\item{rDiscr}{The "Discretized Connectivity Matrix" (delivered by "Classify" -- Classes = c(0,1) only)}

\item{Ref}{The "Reference Matrix"}

\item{Title}{Title of the graph.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{		A message if an error is detected and returns NULL otherwise.
}\if{html}{\out{</div>}}
}
\description{
This function checks the input data for function DrawDiscr.
The parameters are the same as those of DrawDiscr.
}
