% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawGraph.R
\name{CheckInputDataDG}
\alias{CheckInputDataDG}
\title{Checks the input data for function DrawGraph}
\usage{
CheckInputDataDG(Ret, Title, Thr)
}
\arguments{
\item{Ret}{list of informations delivered by "MRARegress" or a square matrix.}

\item{Title}{Title of the graph.}

\item{Thr}{Threshold to determine edge existence}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{		A message if an error is detected and returns NULL otherwise.
}\if{html}{\out{</div>}}
}
\description{
This function checks the input data for function DrawGraph.
The parameters are the same as those of DrawGraph.
}
