% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawGraphML.R
\name{CheckInputDataDGML}
\alias{CheckInputDataDGML}
\title{Checks the input data for function DrawGraphML}
\usage{
CheckInputDataDGML(Ret, File, Thr)
}
\arguments{
\item{Ret}{list of informations delivered by "MRARegress" or a square matrix.}

\item{File}{Name of the file to write.}

\item{Thr}{Threshold to determine edge existence}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{		A message if an error is detected and returns NULL otherwise.
}\if{html}{\out{</div>}}
}
\description{
This function checks the input data for function DrawGraphML.
The parameters are the same as those of DrawGraphML.
}
