% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawHeat.R
\name{CheckInputDataDH}
\alias{CheckInputDataDH}
\title{Checks the input data for function DrawHeat}
\usage{
CheckInputDataDH(
  Ret,
  Absolute,
  Title,
  FontSize,
  Colour,
  Legend,
  DisplayNumbers,
  DecNumbers,
  FileName,
  NoPrint
)
}
\arguments{
\item{Ret}{list of informations delivered by "MRARegress" or square matrix of numbers.}

\item{Absolute}{If TRUE, the heatmap relates to the absolute value of the data.}

\item{Title}{Title of the plot.}

\item{FontSize}{Base fontsize for the plot.}

\item{Colour}{Defines the colour palette.}

\item{Legend}{To determine if legend should be drawn or not.}

\item{DisplayNumbers}{If TRUE, the numeric values are printed in the cells.}

\item{DecNumbers}{The number of decimal places of the numbers displayed in the cells (if any).}

\item{FileName}{If Not NULL, the plot is saved in a .pdf file, whose name is "FileName.pdf"}

\item{NoPrint}{If TRUE, no printings are made. Useful for tests including calls to 'DrawHeat' in a loop.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{		A message if an error is detected and returns NULL otherwise.
}\if{html}{\out{</div>}}
}
\description{
This function checks the input data for function DrawHeat.
The parameters are the same as those of DrawHeat.
}
