% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Score.R
\name{CheckInputDataSC}
\alias{CheckInputDataSC}
\title{Checks the input data for function Score}
\usage{
CheckInputDataSC(Ret, Ref, Classes, Verbose, NoPrint)
}
\arguments{
\item{Ret}{list of informations delivered by "MRARegress" or a matrix.}

\item{Ref}{Matrix of integers or discrete values    The "Reference Matrix".}

\item{Classes}{Vector        A vector showing the classes (for instance: c(0,1))}

\item{Verbose}{If TRUE, additional printings are made. These printings are for internal use only, so they are not documented}

\item{NoPrint}{If TRUE, no printings are made. Useful for tests including calls to 'MRARegress' in a loop.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{		A message if an error is detected and returns NULL otherwise.
}\if{html}{\out{</div>}}
}
\description{
This function checks the input data for function Score.
The parameters are the same as those of Score.
}
