% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawGraphML.R
\name{DrawGraphML}
\alias{DrawGraphML}
\title{"DrawGraphML" :   records an XML file, describing "r" as a graph in GraphML format}
\usage{
DrawGraphML(r)
}
\arguments{
\item{r}{Matrix of numbers        The connectivity matrix (delivered by "MRARegress")}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{File	An XML file, describing "r" as a graph in GraphML format.
}\if{html}{\out{</div>}}
}
\description{
:  records an XML file, describing "r" as a graph in GraphML format.
This file may be used directly by drawing software, like Cytoscape or iGraph.
}
