% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawHeat.R
\name{DrawHeat}
\alias{DrawHeat}
\title{"DrawHeat" :   represents "r" as a "matrix heat chart"}
\usage{
DrawHeat(r)
}
\arguments{
\item{r}{Matrix of numbers        The connectivity matrix (delivered by "MRARegress")}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{No return
Displays the "matrix heat chart" associated with the connectivity matrix.
}\if{html}{\out{</div>}}
}
\description{
: this function represents the Connectivity Matrix "r" as a "matrix heat chart", where the colours are a function of the absolute value of r[i,j]
}
