% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRARegress.R
\name{fcarre}
\alias{fcarre}
\title{Computes a squared value}
\usage{
fcarre(x)
}
\arguments{
\item{x}{The value to square (may be a vector)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{	A vector, with the values squared
}\if{html}{\out{</div>}}
}
\description{
Computes a squared value
}
