% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRARegress.R
\name{fELASTIC}
\alias{fELASTIC}
\title{Computes the "Connectivity Matrix" using ELASTICNET, RIDGE or LASSO Method (library glmnet)}
\usage{
fELASTIC(MatU, MatY, Lbda, Mu, nbN)
}
\arguments{
\item{MatU}{Used to compute linear Regression}

\item{MatY}{Y as a function of U}

\item{Lbda}{Hyper parameter λ of the method  -- If Lbda = NULL, a "best value" of Lbda is computed directly}

\item{Mu}{Hyper parameter μ of the method  -- Mu = 0 : RIDGE, Mu = 1 : LASSO, else (between 0 and 1) : Elastic Net.}

\item{nbN}{Number of nodes}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{		NULL if an error is detected or the result ("r[iNode, ]") otherwise (vector, length = nbN)
}\if{html}{\out{</div>}}
}
\description{
Computes the "Connectivity Matrix" using ELASTICNET, RIDGE or LASSO Method (library glmnet)
}
