% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frank_FileNbr.R
\docType{data}
\name{FileNbr}
\alias{FileNbr}
\title{Institut de Recherche en Cancérologie de Montpellier (IRCM)
Cancer Bioinformatics and Systems Biology}
\format{
\subsection{\code{FileNbr}}{

A matrix with 15 rows and 7 columns : Data used to generate FRANK files.
}
}
\source{
<"C:\\Users\\jean-pierre.borg\\IRCM\\These\\Recherche\\Packages\\MRAregress\\data\\Frank_FileNbr.rda">
}
\usage{
FileNbr
}
\description{
A subset of data used by the unit tests of MRARegress
}
\keyword{datasets}
