% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRARegress.R
\name{fRForest}
\alias{fRForest}
\title{Computes the "Connectivity Matrix" using Random Forest Method (library randomForest)}
\usage{
fRForest(MatU, MatY, seqLetters, nbN)
}
\arguments{
\item{MatU}{Used to compute linear Regression}

\item{MatY}{Y as a function of U}

\item{seqLetters}{A sequence of lower case letters "a", "b" .... used to name the columns}

\item{nbN}{Number of nodes}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{		NULL if an error is detected or the result ("r[iNode, ]") otherwise  (vector, length = nbN-1)
}\if{html}{\out{</div>}}
}
\description{
Computes the "Connectivity Matrix" using Random Forest Method (library randomForest)
}
