% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRARegress.R
\name{fSTEP}
\alias{fSTEP}
\title{Computes the "Connectivity Matrix" using STEP Method}
\usage{
fSTEP(MatU, MatY, Meth, nbN, Verbose)
}
\arguments{
\item{MatU}{Used to compute linear Regression}

\item{MatY}{Y as a function of U}

\item{Meth}{Hyper parameter of the method  -- "Fo" => "STEP-Fo", "Ba" => "STEP-Ba", "Bo" => "STEP-Bo"}

\item{nbN}{Number of nodes}

\item{Verbose}{Logical. If TRUE, additional printings are made. These printings are for internal use only, so they are not documented.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{		NULL if an error is detected or the result ("r[iNode, ]") otherwise  (vector, length = nbN-1)
}\if{html}{\out{</div>}}
}
\description{
Computes the "Connectivity Matrix" using STEP Method
}
