% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data1.R
\docType{data}
\name{MapExper}
\alias{MapExper}
\title{Institut de Recherche en Cancérologie de Montpellier (IRCM)
Cancer Bioinformatics and Systems Biology}
\format{
\subsection{\code{MapExper}}{

A matrix with 6 rows and 6 columns (the "Experimental Map"):
MapExper[i,j] = 1 means that the perturbation j perturbs the node i,
MapExper[i,j] = 0 means that the perturbation j does not perturb the node i,
MapExper[i,j] = x means that we do not know if the perturbation j perturbs the node i or not.
}
}
\source{
<"C:\\Users\\jean-pierre.borg\\IRCM\\These\\Recherche\\Packages\\MRAregress\\data-raw\\DATASET.R">
}
\usage{
MapExper
}
\description{
A subset of data used by the first example of the vignette
}
\details{
Report ...
}
\keyword{datasets}
