% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Frank_TF30_TA0_1_R1.R
\docType{data}
\name{Mat_TF30_TA0_1_R1}
\alias{Mat_TF30_TA0_1_R1}
\title{Institut de Recherche en Cancérologie de Montpellier (IRCM)
Cancer Bioinformatics and Systems Biology}
\format{
\subsection{\code{Mat_TF30_TA0_1_R1}}{

A matrix with 60 rows and 30 columns, corresponding to the MatRN matrix got from FRANK (TF=30, TA=0, file 1, 10\% noise).
See the article "Modular response analysis reformulated as a multilinear regression problem" (Borg et al, 2023).

This matrix is for internal use only.
}
}
\source{
<"C:\\Users\\jean-pierre.borg\\IRCM\\These\\Recherche\\Packages\\MRAregress\\data\\Frank_TF30_TA0_1_R1.rda">
}
\usage{
Mat_TF30_TA0_1_R1
}
\description{
A subset of data used by the unit tests of MRARegress
}
\keyword{datasets}
