% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data2.R
\docType{data}
\name{MatExp}
\alias{MatExp}
\title{Institut de Recherche en Cancérologie de Montpellier (IRCM)
Cancer Bioinformatics and Systems Biology}
\format{
\subsection{\code{MatExp}}{

A matrix with 6 rows and 7 columns (the "Expression Matrix"):
MatExp[i,1..nbBase] represents the measurement of node i when no perturbation is applied to the network (here, nbBase = 1),
MatExp[i,j], j>nbBase,  represents the measurement of node i when perturbation j-nbBase is applied to the network.
}
}
\source{
<"C:\\Users\\jean-pierre.borg\\IRCM\\These\\Recherche\\Packages\\MRAregress\\data-raw\\DATASET.R">
}
\usage{
MatExp
}
\description{
A subset of data used by the first example of the vignette
}
\details{
Report ...
}
\keyword{datasets}
