% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data1_MatExp1.R
\docType{data}
\name{MatExp1}
\alias{MatExp1}
\title{Institut de Recherche en Cancérologie de Montpellier (IRCM)
Cancer Bioinformatics and Systems Biology}
\format{
\subsection{\code{MatExp1}}{

A matrix with 6 rows and 7 columns (the "Expression Matrix"), to generate the first example of the vignette.
MAP Kinase network, with 6 nodes.
MatExp1[i,1] 	 = measurement of node i, without any perturbation ("basal" values).
MatExp1[i,2:7] = measurement of node i, perturbed by "Perturbation 1" (col. 2), "Perturbation 2" (col. 3), ..., "Perturbation 6" (col. 7).
}
}
\source{
<"C:\\Users\\jean-pierre.borg\\IRCM\\These\\Recherche\\Packages\\MRAregress\\data-raw\\DATASET.R">
}
\usage{
MatExp1
}
\description{
A subset of data used by the first example of the vignette
}
\keyword{datasets}
