% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatR2MatExp.R
\name{MatR2MatExp}
\alias{MatR2MatExp}
\title{"Utility_MatR2MatExp" :  For internal use only}
\usage{
MatR2MatExp(MatR, nbN, nbK)
}
\arguments{
\item{MatR}{Matrix of numbers Matrix generated according to the description in the article "Modular response analysis reformulated as a multilinear
regression problem" (Borg et al, 2023).}

\item{nbN}{Number              Number of nodes.}

\item{nbK}{Number              Number of perturbations per node (concerning the MatR matrices got from FRANK, nbK = 2 :
The first nbN rows correspond to KO perturbations and the rows nbN+1 .. 2*nbN, to KD50 pertubations).}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{List of matrices	Returns a list with Exp=MatExp, Pert=Perturb.
						  To generate MatExp, we suppose Xi = 1 (all i) and Relative = TRUE (Ri,j = 2*(Xi(P+∆Pj)-Xi(P)) / (Xi(P+∆Pj)+Xi(P)))
}\if{html}{\out{</div>}}
}
\description{
This function is a utility intended to convert the MatR matrices, got from FRANK for example, into the matrices MatExp and Perturb,
necessary to use MRARegress. For internal use only.
}
