% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Matr_Frank_TF30_TA0_1_rcc1.R
\docType{data}
\name{MatrCc}
\alias{MatrCc}
\title{Institut de Recherche en Cancérologie de Montpellier (IRCM)
Cancer Bioinformatics and Systems Biology}
\format{
\subsection{\code{MatrCc}}{

A matrix with 30 rows and 30 columns, corresponding to the solution (MatrCc) of the network generated by FRANK (TF=30, TA=0, file 1, 10\% noise).
See the article "Modular response analysis reformulated as a multilinear regression problem" (Borg et al, 2023).

This matrix is for internal use only : to test the utility "MatR2MatExp".
}
}
\source{
<"C:\\Users\\jean-pierre.borg\\IRCM\\These\\Recherche\\Packages\\MRAregress\\data\\Matr_Frank_TF30_TA0_1_rcc1.rda">
}
\usage{
MatrCc
}
\description{
A subset of data used by the unit tests of MRARegress
}
\keyword{datasets}
