% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_MatrCc_STEP_Bo_10_1.R
\docType{data}
\name{MatrCc_STEP_Bo_10_1}
\alias{MatrCc_STEP_Bo_10_1}
\title{Institut de Recherche en Cancérologie de Montpellier (IRCM)
Cancer Bioinformatics and Systems Biology}
\format{
\subsection{\code{MatrCc_STEP_Bo_10_1}}{

A matrix with 10 rows and 10 columns (the "Connectivity Matrix"), the "Connectivity Matrix" got by STEP-Bo(th) method.
Dream Challenge 4, size=10, Fic.1   ===   The network has 10 nodes, perturbations : KD50 (Knock Down 50\%), KO (Knock Out)
MatrCc_STEP_Bo_10_1[i,j] = the coefficient ri,j.
}
}
\usage{
MatrCc_STEP_Bo_10_1
}
\description{
A subset of data used by the unit tests of MRARegress
}
\keyword{datasets}
