% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_MatrCc_TLR_10_5.R
\docType{data}
\name{MatrCc_TLR_10_5}
\alias{MatrCc_TLR_10_5}
\title{Institut de Recherche en Cancérologie de Montpellier (IRCM)
Cancer Bioinformatics and Systems Biology}
\format{
\subsection{\code{MatrCc_TLR_10_5}}{

A matrix with 10 rows and 10 columns (the "Connectivity Matrix"), the "Connectivity Matrix" got by TLR method (before calling DrawDiscr : TLR = LSE).
Dream Challenge 4, size=10, Fic.5   ===   The network has 10 nodes, perturbations : KD50 (Knock Down 50\%), KO (Knock Out)
MatrCc_TLR_10_5[i,j] = the coefficient ri,j.
}
}
\usage{
MatrCc_TLR_10_5
}
\description{
A subset of data used by the unit tests of MRARegress
}
\keyword{datasets}
