% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_MatR_Frank_TF30_TA0_1_R1.R
\docType{data}
\name{MatRN2}
\alias{MatRN2}
\title{Institut de Recherche en Cancérologie de Montpellier (IRCM)
Cancer Bioinformatics and Systems Biology}
\format{
\subsection{\code{MatRN2}}{

A matrix with 60 rows and 30 columns, corresponding to the MatRN matrix got from FRANK (TF=30, TA=0, file 1, 10\% noise).
See the article "Modular response analysis reformulated as a multilinear regression problem" (Borg et al, 2023).

This matrix is for internal use only : to test the utility "MatR2MatExp".
}
}
\source{
<"C:\\Users\\jean-pierre.borg\\IRCM\\These\\Recherche\\Packages\\MRAregress\\data\\MatR_Frank_TF30_TA0_1_R1.rda">
}
\usage{
MatRN2
}
\description{
A subset of data used by the unit tests of MRARegress
}
\keyword{datasets}
