% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_MatR_Frank_TF300_TA0_1_R1.R
\docType{data}
\name{MatRN300}
\alias{MatRN300}
\title{Institut de Recherche en Cancérologie de Montpellier (IRCM)
Cancer Bioinformatics and Systems Biology}
\format{
\subsection{\code{MatRN300}}{

A matrix with 600 rows and 300 columns, corresponding to the MatRN matrix got from FRANK (TF=300, TA=0, file 1, 10\% noise).
See the article "Modular response analysis reformulated as a multilinear regression problem" (Borg et al, 2023).

This matrix is used to test MRARegress against large networks
}
}
\source{
<"C:\\Users\\jean-pierre.borg\\IRCM\\These\\Recherche\\Packages\\MRAregress\\data\\MatR_Frank_TF300_TA0_1_R1.rda">
}
\usage{
MatRN300
}
\description{
A subset of data used by the unit tests of MRARegress
}
\keyword{datasets}
