% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data3.R
\docType{data}
\name{NodeNames}
\alias{NodeNames}
\title{Institut de Recherche en Cancérologie de Montpellier (IRCM)
Cancer Bioinformatics and Systems Biology}
\format{
\subsection{\code{NodeNames}}{

A vector with 6 values (the name of the 6 nodes)
}
}
\source{
<"C:\\Users\\jean-pierre.borg\\IRCM\\These\\Recherche\\Packages\\MRAregress\\data-raw\\DATASET.R">
}
\usage{
NodeNames
}
\description{
A subset of data used by the first example of the vignette
}
\details{
Report ...
}
\keyword{datasets}
