% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ode2MatExp.R
\name{Ode2MatExp}
\alias{Ode2MatExp}
\alias{MExp}
\title{"Utility_MExp" :  For internal use only}
\usage{
MExp(nbN, Perturbs, P0, func, st)
}
\arguments{
\item{nbN}{Number            Number of nodes.}

\item{Perturbs}{Vector            Set of perturbations. The parameters to perturb are multiplied successively by Perturbs[i]
There are no replicates.}

\item{P0}{Vector            Initial value of the parameters to perturb (for instance : initial value of VS1,VS2,VS3 and VS4).}

\item{func}{Function        Function of P (parameters) and X (expression values to compute)}

\item{st}{Vector            Multiroot starts from these values, to solve the system of ODE.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{List of matrices	Returns a list with Exp=MatExp, Pert=Perturb.
}\if{html}{\out{</div>}}
}
\description{
This function is an utility intended to compute the matrices MatExp and Perturb,
necessary to use MRARegress. For internal use only.
Computation of MatExp is done directly from the equation of system dynamics.
}
