% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ode2MatExp2.R
\name{Ode2MatExp2}
\alias{Ode2MatExp2}
\alias{MExp2}
\title{"Utility_MExp2" :  For internal use only}
\usage{
MExp2(nbN, MapExper, Perturb, P0, func, st)
}
\arguments{
\item{nbN}{Number            Number of nodes.}

\item{MapExper}{Matrix            The matrix used by MRARegress
nbM rows (number of parameters), nbPc columns (number of conducted perturbations). See description in "MRARegress.R"}

\item{Perturb}{Number            Amplitude of the perturbation.}

\item{P0}{Vector            Initial value of the parameters to perturb (for instance : initial value of VS1,VS2,VS3 and VS4).}

\item{func}{Function        Function of P (parameters) and X (expression values to compute)}

\item{st}{Vector            Multiroot starts from these values, to solve the system of ODE.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{List of matrices	Returns a list with Exp=MatExp, Pert=Perturb.
}\if{html}{\out{</div>}}
}
\description{
This function is an utility intended to compute the matrices MatExp and Perturb,
necessary to use MRARegress. For internal use only.
Computation of MatExp is done directly from the equation of system dynamics.
Only one perturbation level is allowed, but perturbations may act on many parameters (H6 FALSE)
}
