% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Pat3_O1.R
\docType{data}
\name{Pat3nds_O1}
\alias{Pat3nds_O1}
\title{Institut de Recherche en Cancérologie de Montpellier (IRCM)
Cancer Bioinformatics and Systems Biology}
\format{
\subsection{\code{Pat3nds_O1}}{

A matrix with 3 rows and 3 columns : the "Exact Connectivity Matrix".
Three nodes network (pRaF, ppMEK, ppERK), described in the article "Impact of measurement noise, experimental design,
and estimation methods on Modular Response -- Analysis based network reconstruction" C. Thomaseth et al (Scientific Reports, 2018).
Pat3nds_O1[i,j] = the exact coefficient ri,j, computed with the perturbations (10\%) corresponding to the experimental design described
by P. Ravel in "Présentation_Patrice.pptx" :
}
}
\usage{
Pat3nds_O1
}
\description{
A subset of data used by the unit tests of MRARegress
}
\keyword{datasets}
