% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data4_Perturb2.R
\docType{data}
\name{Perturb2}
\alias{Perturb2}
\title{Institut de Recherche en Cancérologie de Montpellier (IRCM)
Cancer Bioinformatics and Systems Biology}
\format{
\subsection{\code{Perturb2}}{

A vector(length=21) : the "Perturbations Name" for Example 2 of the Vignette
From Dream Challenge 4 (InSilico_10_1 : "basal" values + 2 perturbations : Knock Down and Knock Out).
Perturb2[1] 		= "Base"			Designates the basal column of MatExp2
Perturb2[2 .. 21] = "Pi -> Ni"		Designates the perturbations and the nodes they act upon
As MatExp2 has 10 rows and 21 columns. Perturb is mandatory.
}
}
\source{
<"C:\\Users\\jean-pierre.borg\\IRCM\\These\\Recherche\\Packages\\MRAregress\\data-raw\\DATASET.R">
}
\usage{
Perturb2
}
\description{
A subset of data used by the second example of the vignette
}
\keyword{datasets}
