% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Res_4.R
\docType{data}
\name{Res_4}
\alias{Res_4}
\title{Institut de Recherche en Cancérologie de Montpellier (IRCM)
Cancer Bioinformatics and Systems Biology}
\format{
\subsection{\code{Res_4}}{

A matrix with 4 rows and 4 columns : the "Connectivity Matrix", coefficients got by Order 2 method and extracted form ResO2_4.
Four nodes network, described in the article "Inferring dynamic architecture of cellular networks using time series of gene expression,
protein and metabolite data" Eduardo Sontag et al, 2004
}
}
\usage{
Res_4
}
\description{
A subset of data used by the unit tests of MRARegress
}
\keyword{datasets}
