% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_ResO2_3.R
\docType{data}
\name{ResO2_3}
\alias{ResO2_3}
\title{Institut de Recherche en Cancérologie de Montpellier (IRCM)
Cancer Bioinformatics and Systems Biology}
\format{
\subsection{\code{ResO2_3}}{

A matrix with 3 rows and 5 columns : the "Order 2 Connectivity Matrix", coefficients got by Order 2 method.
Three nodes network (pRaF, ppMEK, ppERK), described in the article "Impact of measurement noise, experimental design,
and estimation methods on Modular Response -- Analysis based network reconstruction" C. Thomaseth et al (Scientific Reports, 2018).
ResO2_3[i,j] :
row i : concerns node i
col j : j in {1,2} : coefficients corresponding to the "linear part" (ie. ri,j with j != i	--		ri,j * DeltaXi,j)
j in {3,4} : coefficients corresponding to the "quadratic part" (ie. si,j,j with j != i --	si,j,j * (DeltaXi,j)^2
j in {5}   : coefficient  corresponding to the "product part" (ie. si,j,k, with j != i, k != i, k > j --  si,j,k * (DeltaXi,j)*DeltaXi,k)
(Example, for i=1 : s2,2,3)
}
}
\usage{
ResO2_3
}
\description{
A subset of data used by the unit tests of MRARegress
}
\keyword{datasets}
