% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_ResO2_4.R
\docType{data}
\name{ResO2_4}
\alias{ResO2_4}
\title{Institut de Recherche en Cancérologie de Montpellier (IRCM)
Cancer Bioinformatics and Systems Biology}
\format{
\subsection{\code{ResO2_4}}{

A matrix with 4 rows and 9 columns : the "Order 2 Connectivity Matrix", coefficients got by Order 2 method.
Four nodes network, described in the article "Inferring dynamic architecture of cellular networks using time series of gene expression,
protein and metabolite data" Eduardo Sontag et al, 2004
ResO2_4[i,j] :
row i : concerns node i
col j : j in {1,3} : coefficients corresponding to the "linear part" (ie. ri,j with j != i	--		ri,j * DeltaXi,j)
j in {4,6} : coefficients corresponding to the "quadratic part" (ie. si,j,j with j != i --	si,j,j * (DeltaXi,j)^2
j in {7,9} : coefficient  corresponding to the "product part" (ie. si,j,k, with j != i, k != i, k > j --  si,j,k * (DeltaXi,j)*DeltaXi,k)
(Example for i=1 : s1,2,3, s1,2,4, s1,3,4)
}
}
\usage{
ResO2_4
}
\description{
A subset of data used by the unit tests of MRARegress
}
\keyword{datasets}
