% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RocPR.R
\name{RocPR}
\alias{RocPR}
\title{"RocPR" :   displays performance curves when a reference is available.}
\usage{
RocPR(r, Ref, Method = "TLR")
}
\arguments{
\item{r}{Matrix of numbers        The connectivity matrix (delivered by "MRARegress")}

\item{Ref}{Matrix of numbers        The "Reference Matrix"}

\item{Method}{String                    The method to use to compute the connectivity matrix. Method to choose among the available methods:
- TLR: a simple Multiple Linear Regression method which minimizes the Least Square Error (default value),
- LASSO: a shrinkage estimator,
- STEP: a variable selection scheme,
- ARACNE: a method based on Mutual Information,
- CLR: another method based on Mutual Information,
- RMNET: another method based on Mutual Information,
- Random Forest: a method used in Machine Learning.#'}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{	List		A list containing the ROC curve, the PR curve, the areas under these curves (AuROC and AuPR), the p-value of these areas.
}\if{html}{\out{</div>}}
}
\description{
: This function displays the ROC curve, the PR curve, the areas under these curves (AuROC and AuPR), the p-value of these areas, when a reference is available.
For "Classes = c(0,1)" only.
}
