% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRARegress.R
\name{sequence_letters}
\alias{sequence_letters}
\title{Computes a sequence of letters : "a", "b", "c", ...., "z", "aa", "ab", .... , "az", "ba", "bb" etc ....}
\usage{
sequence_letters(n)
}
\arguments{
\item{n}{The number of elements of the sequence}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{	A vector, length n, containing the sequence of letters
}\if{html}{\out{</div>}}
}
\description{
Computes a sequence of letters : "a", "b", "c", ...., "z", "aa", "ab", .... , "az", "ba", "bb" etc ....
}
