% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Solution.R
\docType{data}
\name{Solution}
\alias{Solution}
\title{Institut de Recherche en Cancérologie de Montpellier (IRCM)
Cancer Bioinformatics and Systems Biology}
\format{
\subsection{\code{Solution}}{

The "Solution" is delivered by Dream Challenge (DC4 - InSilico_10_1 : real network). It's a matrix with 10 rows and 10 columns.
Solution[i,j] = 1 means that node "j" has an impact on node "i". If this is not the case, Solution[i,j] = 0.
}
}
\source{
<"C:\\Users\\jean-pierre.borg\\IRCM\\These\\Recherche\\Packages\\MRAregress\\data-raw\\DATASET.R">
}
\usage{
Solution
}
\description{
A subset of data used by the second example of the vignette
}
\keyword{datasets}
