% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Sol3G.R
\docType{data}
\name{Solution_3G}
\alias{Solution_3G}
\title{Institut de Recherche en Cancérologie de Montpellier (IRCM)
Cancer Bioinformatics and Systems Biology}
\format{
\subsection{\code{Solution_3G}}{

A matrix with 3 rows and 3 columns : the "Exact Connectivity Matrix", coefficients computed directly from ODE.
Three nodes network (pRaF, ppMEK, ppERK), described in the article "Impact of measurement noise, experimental design,
and estimation methods on Modular Response -- Analysis based network reconstruction" C. Thomaseth et al (Scientific Reports, 2018).
Solution_3G[i,j] = the exact coefficient ri,j.
}
}
\usage{
Solution_3G
}
\description{
A subset of data used by the unit tests of MRARegress
}
\keyword{datasets}
