% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Sol3L.R
\docType{data}
\name{Solution_3L}
\alias{Solution_3L}
\title{Institut de Recherche en Cancérologie de Montpellier (IRCM)
Cancer Bioinformatics and Systems Biology}
\format{
\subsection{\code{Solution_3L}}{

A matrix with 3 rows and 3 columns : the "Exact Connectivity Matrix", coefficients computed directly from ODE.
Linear three nodes network.  Created according to (Alon, 2006).
Solution_3L[i,j] = the exact coefficient ri,j.
}
}
\usage{
Solution_3L
}
\description{
A subset of data used by the unit tests of MRARegress
}
\keyword{datasets}
