% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Sol4G.R
\docType{data}
\name{Solution_4G}
\alias{Solution_4G}
\title{Institut de Recherche en Cancérologie de Montpellier (IRCM)
Cancer Bioinformatics and Systems Biology}
\format{
\subsection{\code{Solution_4G}}{

A matrix with 4 rows and 4 columns : the "Exact Connectivity Matrix", coefficients computed directly from ODE.
Four nodes network, described in the article "Inferring dynamic architecture of cellular networks using time series of gene expression,
protein and metabolite data" Eduardo Sontag et al, 2004
Solution_4G[i,j] = the exact coefficient ri,j.
}
}
\usage{
Solution_4G
}
\description{
A subset of data used by the unit tests of MRARegress
}
\keyword{datasets}
