% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Sol6K.R
\docType{data}
\name{Solution_6K}
\alias{Solution_6K}
\title{Institut de Recherche en Cancérologie de Montpellier (IRCM)
Cancer Bioinformatics and Systems Biology}
\format{
\subsection{\code{Solution_6K}}{

A matrix with 6 rows and 6 columns : the "Exact Connectivity Matrix", coefficients computed directly from ODE.
Six nodes network "Kinases MKK, MKKK et MAPK + bi-phosphorylated forms", described in the articles "Inferring dynamic architecture of cellular
networks using time series of gene expression, protein and metabolite data" Eduardo Sontag et al, 2004 and "Untangling the wires:
A strategy to trace functional interactions in signaling and gene networks" Boris N. Kholodenko et al, PNAS 2002
Solution_6K[i,j] = the exact coefficient ri,j.
}
}
\usage{
Solution_6K
}
\description{
A subset of data used by the unit tests of MRARegress
}
\keyword{datasets}
