% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classify.R
\name{Thresh}
\alias{Thresh}
\title{Finds the class of a value, according thresholds}
\usage{
Thresh(x, Seuils)
}
\arguments{
\item{x}{The value to classify (may be a vector)}

\item{Seuils}{We compare x with these values}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{	A vector, with the class of the value
}\if{html}{\out{</div>}}
}
\description{
Finds the class of a value, according thresholds
}
