% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawGraph.R
\name{DrawGraph}
\alias{DrawGraph}
\title{"DrawGraph" :   displays "r" as a graph}
\usage{
DrawGraph(r)
}
\arguments{
\item{r}{Matrix of numbers        The connectivity matrix (delivered by "MRARegress")}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{No return
Displays the graph associated with the connectivity matrix.
}\if{html}{\out{</div>}}
}
\description{
: this function represents the Connectivity Matrix "r" as a graph where the nodes are the genes (for instance) \
and an edge A $\\rightarrow$ B reflects the action of node A on node B (amplification), while a dashed line (with a "T" arrow  betwwen A and B) reflects an inhibition of node B by node A.
}
